Add postgrex adapter for subscription
