+++ lib/event_store/storage_adapters/ecto/repo.ex
@@ -1,0 +1,3 @@
+defmodule EventStore.StorageAdapters.Ecto.Repo do
+  use Ecto.Repo, otp_app: :eventstore
+end
