+++ lib/event_store/storage_adapters/manager.ex
@@ -1,0 +1,5 @@
+defmodule EventStore.StorageAdapters.Manager do
+  def storage_adapter() do
+    Application.get_env(:eventstore, :storage_adapter)
+  end 
+end
